cloud_app/ws/requirements.txt
Flask
Flask-CORS
flask_sqlalchemy
Flask-Migrate
mysql-connector-python
pandas
werkzeug
PyJWT
pymysql
cryptography
python-dotenv
gunicorn
requests
pytest
bcrypt
scikit-learn